+++ NEWS.md
@@ -5,6 +5,7 @@
 - Optimized `get_R_package_usage()` function: 
   - you can now pass a vector of packages names (new `packages` parameter replacing old `package_name`) ([#494](https://github.com/r-world-devs/GitStats/issues/494)),
   - on the other hand, output of the function has been limited to contain only most necessary data (removing all repository stats), making thus process of obtaining package usage faster ([#474](https://github.com/r-world-devs/GitStats/issues/474)).
+  - new `split_output` parameter has been added - when set to `TRUE` a list with tibbles (every element of the list for every package) instead of one tibble is returned.
 - Added possibility to get repositories for individual users with `get_repos()` ([#492](https://github.com/r-world-devs/GitStats/issues/492)). Earlier this was only possible for GitHub organizations and GitLab groups.
 
 ## Fixes: